==> 02-scripts/010_analysis_paired_include_guon/10_pseudobulk_nsclc.ipynb <==
# 1.NSCLC Sex-stratified SC Analysis

## Data selection, pseudobulk, create input for DESEQ2

import warnings
import numpy as np
import pandas as pd 
import scanpy as sc
#import tensorflow as tf
import matplotlib.pyplot as plt
import matplotlib
import seaborn as sns
#import altair as alt
#import pertpy as pt
#from pandas.api.types import is_categorical_dtype
#from pandas import CategoricalDtype

import decoupler as dc

dc.__version__

path = "/data/projects/2023/LCBiome/nsclc_gender_atlas_tmp"
input_path    = f"{path}/data/local.h5ad"

adata = sc.read_h5ad(input_path) 

adata.obs.dataset.value_counts()

nsclc_chrom = pd.read_csv(f"{path}/out/007_re_analysis/tables/input/adata_var_nsclc_chrom.csv")

adata

adata.obs.disease.value_counts()

adata.obs.tumor_stage.value_counts()

adata.obs.sex.value_counts()

adata.obs.origin.value_counts()

adata.obs.dataset.value_counts()

### 01.SELECT ONLY male, female & tumor, normal_adjacent

adata = adata[(adata.obs["disease"] != "normal") & (adata.obs["disease"] != "chronic obstructive pulmonary disease")] #exclude normal and COPD samples

adata = adata[adata.obs["tumor_stage"].notna()] #exclude samples without tumor stage

adata = adata[(adata.obs["sex"]=="male")| (adata.obs["sex"]=="female")]

adata = adata[(adata.obs["origin"]=="tumor_primary")| (adata.obs["origin"]=="normal_adjacent")]

adata.obs.disease.value_counts()

adata.obs.tumor_stage.value_counts()

adata.obs.sex.value_counts()

adata.obs.origin.value_counts()

adata.obs.dataset.value_counts()

### 02.CELL TPYE RE-CLASIFICATION 

adata.obs.cell_type.value_counts()

adata.obs.cell_type_major.value_counts()

cells_of_interest=['alveolar macrophage', 'neutrophil',
 'macrophage','malignant cell',
 'B cell',
  'regulatory T cell',
'CD4-positive, alpha-beta T cell',
 'CD8-positive, alpha-beta T cell',
 'neutrophils','epithelial cell of lung','multi-ciliated epithelial cell' ]
subset_adata = adata[adata.obs["cell_type"].isin(cells_of_interest)] 


set(subset_adata.obs.cell_type)

set(subset_adata.obs.cell_type_major)

subset_adata.obs["cell_type"]= subset_adata.obs["cell_type"].replace(['epithelial cell of lung','multi-ciliated epithelial cell',], 'epithelial cell')
subset_adata.obs["cell_type"]= subset_adata.obs["cell_type"].replace(['alveolar macrophage'], 'macrophage')
subset_adata.obs["cell_type"]= subset_adata.obs["cell_type"].replace(['CD4-positive, alpha-beta T cell'], 'CD4+ T cell')
subset_adata.obs["cell_type"]= subset_adata.obs["cell_type"].replace(['CD8-positive, alpha-beta T cell'], 'CD8+ T cell')

set(subset_adata.obs.cell_type)

subset_adata

subset_adata.obs.disease.value_counts()

subset_adata.obs.dataset.value_counts()

subset_adata.obs.dataset.value_counts()

## PAIRED SAMPLES REMOVE NSCLC 

subset_adata = subset_adata[(subset_adata.obs["disease"]!="non-small cell lung carcinoma")]

subset_adata.obs.dataset.value_counts()

subset_adata = subset_adata[(subset_adata.obs["disease"]!="squamous cell lung carcinoma")]

subset_adata.obs.dataset.value_counts()

df_grouped = subset_adata.obs.groupby(["origin","donor_id"]).count()

df_grouped.reset_index(inplace= True)

df_grouped_normal = df_grouped[df_grouped["origin"]=="normal_adjacent"]

df_grouped_normal = df_grouped_normal[df_grouped_normal['sample'] != 0]

df_grouped_tumor = df_grouped[df_grouped["origin"]=="tumor_primary"]

df_grouped_tumor = df_grouped_tumor[df_grouped_tumor['sample'] != 0]

list(set(df_grouped_tumor.donor_id))

common_ids =list(set(df_grouped_normal.donor_id)& set(df_grouped_tumor.donor_id))

len(common_ids)

paired_adata = subset_adata[subset_adata.obs['donor_id'].isin(common_ids)].copy()

df = paired_adata.obs#
origin_counts = df.groupby('origin')['donor_id'].nunique().reset_index()
origin_counts.columns = ['origin', 'count']

plt.figure(figsize=(10, 6))
sns.barplot(data=origin_counts, x='origin', y='count', palette='viridis')
plt.xlabel('Origin')
plt.ylabel('Number of Unique Donor IDs')
plt.title('Number of Unique Donor IDs per Origin')
plt.show()

df = paired_adata.obs#
origin_counts = df.groupby('disease')['donor_id'].nunique().reset_index()
origin_counts.columns = ['disease', 'count']

plt.figure(figsize=(10, 6))
sns.barplot(data=origin_counts, x='disease', y='count', palette='viridis')
plt.xlabel('disease')
plt.ylabel('Number of Unique Donor IDs')
plt.title('Number of Unique Donor IDs per disease')
plt.show()

## SUMMARY 
- Filtered for male, female & tumor, normal_adjacent
- Filtered and reclassified cells of interest: 'B cell','CD4+ T cell', 'CD8+ T cell', 'epithelial cell', 'macrophage', 'malignant cell', 'neutrophil', 'regulatory Tcell'
- Filtered to exclude samples coming from NSCLC
- Filtered for PAIRED SAMPLES (82 donor_id that have normal_adjacent and tumor_primary  samples)
- KEEP LUADN AND LUSC

### 04.Compute pseudobulk

# Get pseudo-bulk profile
pdata = dc.get_pseudobulk(
    paired_adata,
    sample_col='sample',
    groups_col='donor_id',
    layer='count',
    min_cells=0,
    min_counts=0
)

#pdata_var_concat =pd.concat([pdata.var, nsclc_chrom],axis=1)

#pdata_var_concat

pdata.obs["age"] = pdata.obs["age"].astype("float64")
pdata.obs["is_primary_data"] = pdata.obs["is_primary_data"].astype(bool)


#pdata.write_h5ad(f"{path}/out/010_analysis_paired_include_guon/tables/input/pdata.h5ad")

#pdata = sc.read_h5ad(f"{path}/out/008_analysis_paired_luad/tables/input/pdata.h5ad")

pdata

dc.plot_psbulk_samples(pdata, groupby=['dataset', 'platform'], figsize=(12, 4))

#pdata = pdata[pdata.obs["dataset"]!="Guo_Zhang_2018"] ## SMART SEQ 

#pdata

#dc.plot_psbulk_samples(pdata, groupby=['dataset', 'platform'], figsize=(12, 4))

dc.plot_filter_by_expr(pdata, group='sample', min_count=10, min_total_count=15)

dc.plot_filter_by_expr(pdata, group='sample', large_n=100,min_prop=0.9, min_count=5000, min_total_count=10000)

I set this threshold for the DS analysis to avoid havinh zero-inflated data

# Obtain genes that pass the thresholds 
genes = dc.filter_by_expr(pdata, group='sample', min_count=5000, min_total_count=10000)

# Filter by these genes
pdata = pdata[:, genes].copy()
pdata

pdata.layers["counts"] = pdata.X

scales_counts = sc.pp.normalize_total(pdata, target_sum=None, inplace=False)
# log1p transform
pdata.layers["log1p_norm"] = sc.pp.log1p(scales_counts["X"], copy=True)

pdata

### Samplesheet
samplesheet = pdata.obs.copy()
samplesheet["sample"]=samplesheet.index
samplesheet.reset_index(inplace=True)

samplesheet.to_csv(f"{path}/out/010_analysis_paired_include_guon/tables/input/samplesheet.csv", index=False)
samplesheet.to_csv(f"{path}/out/010_analysis_paired_include_guon/tables/input/samplesheet.tsv", index=False, sep = "\t")

bulk_df = pdata.to_df().T

### Raw counts
bulk_df = pdata.to_df().T
#bulk_df["gene_id"]=bulk_df.index
bulk_df.columns = bulk_df.columns.str.replace("[.-]", "_").str.replace(" ", "_")

bulk_df.to_csv(f"{path}/out/010_analysis_paired_include_guon/tables/input/counts.csv")    

### log1p_norm counts
pdata_log1p_norm = pd.DataFrame(pdata.layers["log1p_norm"], 
    pdata.to_df().index, 
    pdata.to_df().columns)
bulk_df = pdata_log1p_norm.T
#bulk_df["gene_id"]=bulk_df.index
bulk_df.columns = bulk_df.columns.str.replace("[.-]", "_").str.replace(" ", "_")
bulk_df.to_csv(f"{path}/out/010_analysis_paired_include_guon/tables/input/log1p_norm_counts.csv")    

### 0.4 Create counts and samplesheet independent for tumor and normal for DESEQ2 

pdata_normal = pdata[pdata.obs["origin"]=="normal_adjacent"]
pdata_tumor = pdata[pdata.obs["origin"]=="tumor_primary"]

pdata_normal

pdata_list = [pdata_normal,pdata_tumor]
pdata_list_name = ["normal","tumor"]

for index, dataframe in enumerate(pdata_list):
    print(pdata_list_name[index])
    # Perform the operations
    samplesheet = dataframe.obs.copy()
    samplesheet["sample"] = samplesheet.index
    samplesheet.reset_index(inplace=True)
  
    samplesheet.to_csv(f"{path}/out/010_analysis_paired_include_guon/tables/input/samplesheet_{pdata_list_name[index]}.csv", index=False)
    samplesheet.to_csv(f"{path}/out/010_analysis_paired_include_guon/tables/input/samplesheet_{pdata_list_name[index]}.tsv", index=False, sep="\t")


    ### Raw counts
    bulk_df = dataframe.to_df().T
    #bulk_df["gene_id"]=bulk_df.index
    bulk_df.columns = bulk_df.columns.str.replace("[.-]", "_").str.replace(" ", "_")
    bulk_df.to_csv(f"{path}/out/010_analysis_paired_include_guon/tables/input/counts_{pdata_list_name[index]}.csv")  
    ### log1p_norm counts
    pdata_log1p_norm = pd.DataFrame(dataframe.layers["log1p_norm"], 
        dataframe.to_df().index, 
        dataframe.to_df().columns)
    bulk_df = pdata_log1p_norm.T
    #bulk_df["gene_id"]=bulk_df.index
    bulk_df.columns = bulk_df.columns.str.replace("[.-]", "_").str.replace(" ", "_")
    bulk_df.to_csv(f"{path}/out/010_analysis_paired_include_guon/tables/input/log1p_norm_counts_{pdata_list_name[index]}.csv")    

## SUMMARY 
- Filtered for male, female & tumor, normal_adjacent
- Filtered and reclassified cells of interest: 'B cell','CD4+ T cell', 'CD8+ T cell', 'epithelial cell', 'macrophage', 'malignant cell', 'neutrophil', 'regulatory Tcell'
- Filtered to exclude samples coming from NSCLC
- Filtered for PAIRED SAMPLES (82 donor_id that have normal_adjacent and tumor_primary  samples)
